requirements.txt
requests
PyMuPDF<1.24.0
Pillow
google-generativeai
python-dotenv